requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-vision==3.4.5
google-cloud-videointelligence==2.11.4
google-cloud-secret-manager==2.16.4
aiohttp==3.9.1
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
